requirements.txt
pyautogen
openai
numpy
pandas
python-dotenv